+++ esp32_bt_mic/src/bt_init.h
@@ -51,13 +51,18 @@ void bt_audio_set_active(bool active);
 void bt_hfp_hf_wake_acl(void);
 
 /**
- * @brief Activate Classic BT HFP (make discoverable/connectable, trigger auto-connect)
+ * @brief Activate Classic BT HFP SCO channel (PTT active)
  */
 void bt_classic_activate(void);
 
 /**
- * @brief Deactivate Classic BT HFP (make non-discoverable/non-connectable, disconnect if connected)
+ * @brief Deactivate Classic BT HFP SCO channel (PTT released)
  */
 void bt_classic_deactivate(void);
+
+/**
+ * @brief Get whether PTT mode is currently activated
+ */
+bool bt_classic_is_ptt_activated(void);
 
 #endif /* __BT_INIT_H__ */